Fix Config TLS plugin default verb
