+++ include/osquery/config.h
@@ -322,6 +322,7 @@ class Config : private boost::noncopyable {
   FRIEND_TEST(SchedulerTests, test_monitor);
   FRIEND_TEST(SchedulerTests, test_config_results_purge);
   FRIEND_TEST(EventsTests, test_event_subscriber_configure);
+  FRIEND_TEST(TLSConfigTests, test_retrieve_config);
 };
 
 /**
